+++ restaurant/v1/serializers.py
@@ -10,6 +10,7 @@ from rest_framework.relations import SlugRelatedField
 
 import geodata.models as geodata_models
 import restaurant.models as models
+from restaurant.services import DishClassifier
 
 info_logger = logging.getLogger('info_logger')
 
@@ -68,6 +69,8 @@ class RestaurantSerializer(serializers.ModelSerializer):
             'menu_url': validated_data.get('menu_url'),
         }
         restaurant, created = models.Restaurant.objects.update_or_create(defaults=validated_data, **lookup)
+        if created:
+            restaurant.save_nearest_stations()
         return restaurant
 
     @transaction.atomic
@@ -117,6 +120,7 @@ class DishSerializer(serializers.ModelSerializer):
         rep['tags'] = TagSerializer(instance.tags.all(), many=True, context=self.context).data
         return rep
 
+    @transaction.atomic
     def create(self, validated_data: dict) -> models.Dish:
         tags = validated_data.pop('tags', [])
 
@@ -130,6 +134,9 @@ class DishSerializer(serializers.ModelSerializer):
 
         if tags:
             dish.tags.set(tags)
+        else:
+            tag_names = DishClassifier(dish).classify_dish()
+            dish.create_tags(tag_names)
 
         return dish