+++ CHANGELOG.md
@@ -15,6 +15,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [1.3.0] - 2025-12-04
+
+### Added
+
+- Dish classifier service ([#31](https://github.com/kooznitsa/sampo/pull/31)).
+- Dish tags generation on save ([#32](https://github.com/kooznitsa/sampo/pull/32)).
+- Dish tags generation commands ([#33](https://github.com/kooznitsa/sampo/pull/33)).
+- Admin methods to create restaurant nearest stations ([#34](https://github.com/kooznitsa/sampo/pull/34)).
+- Filter list of dishes by tags ([#36](https://github.com/kooznitsa/sampo/pull/36)).
+- Production settings ([#38](https://github.com/kooznitsa/sampo/pull/38)).
+
+### Removed
+
+- Permission limitations for GET requests ([#40](https://github.com/kooznitsa/sampo/pull/40)).
+
 ## [1.2.1] - 2025-11-18
 
 ### Added